pinctrl: sunxi: add support for known suniv SoCs.
The suniv chip family of Allwinner (which seems to share the same die) has a
pin controller like other SoCs from Allwinner.

Although some pins are not routed out of the packages with SIP DDR1, they are
also used and internally tied to the DRAM. So they also uses the driver
with all pins.

Add support for these SoCs' pin controller.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>